feat: initial commit w/ attestation and cosign
This adds a `FROM scratch` Docker image that only copies `cacerts` from
a Debian base and pushes it.

I wanted this so I had a secure container that could be used instead of
doing this myself on each project of mine (mainly Go). I've integrated
`cosign` and attestation as security features for others looking to use
this.